DS-16678: remove message text and subject from validation request objects
